Changed room type generation
mapGen.py:
Secret room still depends on passing the comparison to secretChance
parameter, but now there is a 75% chance that the secret room will be a
Hidden Room. 25% chance to be an NPC room, who will ask the players to
complete a quest for him.

room.py:
Added support for the two types of secret room.